+++ src/main/java/org/opentripplanner/graph_builder/module/osm/ElevatorProcessor.java
@@ -84,10 +84,15 @@ class ElevatorProcessor {
       for (OSMLevel level : levels) {
         // get the node to build the elevator out from
         OsmVertex sourceVertex = vertices.get(level);
-        String sourceVertexLabel = sourceVertex.getLabel();
         String levelName = level.longName;
 
-        createElevatorVertices(graph, onboardVertices, sourceVertex, sourceVertexLabel, levelName);
+        createElevatorVertices(
+          graph,
+          onboardVertices,
+          sourceVertex,
+          sourceVertex.getLabelString(),
+          levelName
+        );
       }
       int travelTime = parseDuration(node).orElse(-1);
 
@@ -121,7 +126,7 @@ class ElevatorProcessor {
       for (int i = 0; i < nodes.size(); i++) {
         Long node = nodes.get(i);
         var sourceVertex = vertexGenerator.intersectionNodes().get(node);
-        String sourceVertexLabel = sourceVertex.getLabel();
+        String sourceVertexLabel = sourceVertex.getLabelString();
         String levelName = elevatorWay.getId() + " / " + i;
         createElevatorVertices(
           graph,
@@ -151,24 +156,20 @@ class ElevatorProcessor {
     Graph graph,
     ArrayList<Vertex> onboardVertices,
     IntersectionVertex sourceVertex,
-    String sourceVertexLabel,
+    String label,
     String levelName
   ) {
     var factory = new VertexFactory(graph);
     ElevatorOffboardVertex offboardVertex = factory.elevatorOffboard(
       sourceVertex,
-      sourceVertexLabel,
+      label,
       levelName
     );
 
     FreeEdge.createFreeEdge(sourceVertex, offboardVertex);
     FreeEdge.createFreeEdge(offboardVertex, sourceVertex);
 
-    ElevatorOnboardVertex onboardVertex = factory.elevatorOnboard(
-      sourceVertex,
-      sourceVertexLabel,
-      levelName
-    );
+    ElevatorOnboardVertex onboardVertex = factory.elevatorOnboard(sourceVertex, label, levelName);
 
     ElevatorBoardEdge.createElevatorBoardEdge(offboardVertex, onboardVertex);
     ElevatorAlightEdge.createElevatorAlightEdge(