+++ src/main/java/org/opentripplanner/street/model/vertex/StationElementVertex.java
@@ -2,12 +2,21 @@ package org.opentripplanner.street.model.vertex;
 
 import javax.annotation.Nonnull;
 import org.opentripplanner.framework.i18n.I18NString;
+import org.opentripplanner.transit.model.framework.FeedScopedId;
 import org.opentripplanner.transit.model.site.StationElement;
 
 public abstract class StationElementVertex extends Vertex {
 
-  protected StationElementVertex(String label, double x, double y, I18NString name) {
-    super(label, x, y, name);
+  private final FeedScopedId id;
+
+  protected StationElementVertex(FeedScopedId id, double x, double y, I18NString name) {
+    super(x, y, name);
+    this.id = id;
+  }
+
+  @Override
+  public final VertexLabel getLabel() {
+    return VertexLabel.feedScopedId(id);
   }
 
   /** Get the corresponding StationElement */