+++ src/test/java/org/opentripplanner/street/model/edge/PathwayEdgeTest.java
@@ -3,6 +3,7 @@ package org.opentripplanner.street.model.edge;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.opentripplanner.street.model._data.StreetModelForTest.intersectionVertex;
 
 import java.util.stream.Stream;
 import org.junit.jupiter.api.Test;
@@ -11,7 +12,7 @@ import org.junit.jupiter.params.provider.Arguments;
 import org.opentripplanner.framework.i18n.NonLocalizedString;
 import org.opentripplanner.routing.api.request.StreetMode;
 import org.opentripplanner.routing.api.request.preference.WheelchairPreferences;
-import org.opentripplanner.street.model.vertex.SimpleVertex;
+import org.opentripplanner.street.model._data.StreetModelForTest;
 import org.opentripplanner.street.model.vertex.Vertex;
 import org.opentripplanner.street.search.request.StreetSearchRequest;
 import org.opentripplanner.street.search.state.State;
@@ -20,8 +21,8 @@ import org.opentripplanner.transit.model.site.PathwayMode;
 
 class PathwayEdgeTest {
 
-  Vertex from = new SimpleVertex("A", 10, 10);
-  Vertex to = new SimpleVertex("B", 10.001, 10.001);
+  Vertex from = intersectionVertex(10, 10);
+  Vertex to = intersectionVertex(10.001, 10.001);
 
   @Test
   void zeroLength() {