+++ src/test/java/org/opentripplanner/street/model/vertex/SimpleVertex.java
@@ -4,7 +4,15 @@ import org.opentripplanner.framework.i18n.NonLocalizedString;
 
 public class SimpleVertex extends StreetVertex {
 
+  private final String label;
+
   public SimpleVertex(String label, double lat, double lon) {
-    super(label, lon, lat, new NonLocalizedString(label));
+    super(lon, lat, new NonLocalizedString(label));
+    this.label = label;
+  }
+
+  @Override
+  public VertexLabel getLabel() {
+    return VertexLabel.string(label);
   }
 }